+++ connection.impl/src/main/java/com/ericsson/bss/cassandra/ecchronos/connection/impl/builders/DistributedNativeBuilder.java
@@ -244,10 +244,15 @@ public class DistributedNativeBuilder
         LOG.info("Requesting Nodes List");
         List<Node> nodesList = createNodesList(session);
         LOG.info("Nodes list was created with success");
-        return new DistributedNativeConnectionProviderImpl(session, nodesList);
+        return new DistributedNativeConnectionProviderImpl(session, nodesList, this);
     }
 
-    private List<Node> createNodesList(final CqlSession session)
+    /**
+     * Creates a list of nodes based on the connection type, reads the node list from the database.
+     * @param session the connection information to the database
+     * @return list of nodes
+     */
+    public List<Node> createNodesList(final CqlSession session)
     {
         List<Node> tmpNodeList = new ArrayList<>();
         switch (myType)
@@ -268,6 +273,49 @@ public class DistributedNativeBuilder
         }
         return tmpNodeList;
     }
+
+    /**
+     * Checks the node is on the list of specified dc's/racks/nodes.
+     * @param node
+     * @return
+     */
+    public Boolean confirmNodeValid(final Node node)
+    {
+        switch (myType)
+        {
+            case datacenterAware:
+                return confirmDatacenterNodeValid(node, myDatacenterAware);
+            case rackAware:
+                return confirmRackNodeValid(node, myRackAware);
+            case hostAware:
+                return confirmHostNodeValid(node, myHostAware);
+
+            default:
+        }
+        return false;
+    }
+
+    private Boolean confirmDatacenterNodeValid(final Node node, final List<String> datacenterNames)
+    {
+        return (datacenterNames.contains(node.getDatacenter()));
+    }
+
+    private Boolean confirmRackNodeValid(final Node node, final List<Map<String, String>> rackInfo)
+    {
+        Set<Map<String, String>> racksInfoSet = new HashSet<>(rackInfo);
+        Map<String, String> tmpRackInfo = new HashMap<>();
+        tmpRackInfo.put("datacenterName", node.getDatacenter());
+        tmpRackInfo.put("rackName", node.getRack());
+        return (racksInfoSet.contains(tmpRackInfo));
+    }
+
+    private Boolean confirmHostNodeValid(final Node node, final List<InetSocketAddress> hostsInfo)
+    {
+        Set<InetSocketAddress> hostsInfoSet = new HashSet<>(hostsInfo);
+
+        InetSocketAddress tmpAddress = (InetSocketAddress) node.getEndPoint().resolve();
+        return (hostsInfoSet.contains(tmpAddress));
+     }
 
     private CqlSession createSession(final DistributedNativeBuilder builder)
     {